requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0

alembic==1.13.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

segno 

# Enhanced SmartAdd dependencies
opencv-python==4.8.1.78
pyzbar==0.1.9
numpy==1.24.3
google-generativeai==0.3.2
pillow==10.1.0
qrcode==7.4.2
